requirements.txt
requests
websockets
aiohttp
numpy
pytz
pandas_ta